放大镜

+++ IBY/IBY/Core/category/UIImage+Resize.m
@@ -79,7 +79,7 @@
             break;
             
         default:
-            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %d", contentMode];
+            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %d", (int)contentMode];
     }
     
     CGSize newSize = CGSizeMake(self.size.width * ratio, self.size.height * ratio);
@@ -292,42 +292,5 @@ static inline CGFloat DegreesToRadians(CGFloat degrees)
     
     return transform;
 }
-//NSUInteger alphaOffset(NSUInteger x, NSUInteger y, NSUInteger w){return y * w * 4 + x * 4 + 0;}
-//
-//unsigned char *getBitmapFromImage (UIImage *image)
-//{
-//    CGColorSpaceRef colorSpace = CGColorSpaceCreateDeviceRGB();
-//    if (colorSpace == NULL)
-//    {
-//        fprintf(stderr, "Error allocating color space\n");
-//        return NULL;
-//    }
-//    
-//    CGSize size = image.size;
-//    // void *bitmapData = malloc(size.width * size.height * 4);
-//    unsigned char *bitmapData = calloc(size.width * size.height * 4, 1); // Courtesy of Dirk. Thanks!
-//    if (bitmapData == NULL)
-//    {
-//        fprintf (stderr, "Error: Memory not allocated!");
-//        CGColorSpaceRelease(colorSpace);
-//        return NULL;
-//    }
-//    
-//    CGContextRef context = CGBitmapContextCreate (bitmapData, size.width, size.height, 8, size.width * 4, colorSpace, kCGImageAlphaPremultipliedFirst);
-//    CGColorSpaceRelease(colorSpace );
-//    if (context == NULL)
-//    {
-//        fprintf (stderr, "Error: Context not created!");
-//        free (bitmapData);
-//        return NULL;
-//    }
-//    
-//    CGRect rect = CGRectMake(0.0f, 0.0f, size.width, size.height);
-//    CGContextDrawImage(context, rect, image.CGImage);
-//    unsigned char *data = CGBitmapContextGetData(context);
-//    CGContextRelease(context);
-//    
-//    return data;
-//}
 
 @end
